requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlmodel==0.0.19
pyyaml==6.0.1
python-multipart==0.0.9
pytest==8.2.0
PyMuPDF==1.27.2.2
requests==2.33.1
